+++ .gitignore
@@ -23,3 +23,4 @@ dist-ssr
 *.sw?
 .env*.local
 .env
+tsconfig.tsbuildinfo
