+++ index.html
@@ -1,13 +1,13 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="UTF-8"/>
-    <link href="/vite.svg" rel="icon" type="image/svg+xml"/>
-    <meta content="width=device-width, initial-scale=1.0" name="viewport"/>
-    <title>TDL</title>
+    <link rel="icon" type="image/svg+xml" href="/vite.svg"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+    <title>Todo App - Feature Sliced Design</title>
 </head>
 <body>
 <div id="root"></div>
-<script src="/src/main.tsx" type="module"></script>
+<script type="module" src="/src/app/index/entry.tsx"></script>
 </body>
 </html>
