+++ package.json
@@ -1,16 +1,20 @@
 {
   "name": "react-ts",
+  "description": "React TypeScript Todo application with Supabase integration",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.1.0",
   "type": "module",
+  "engines": {
+    "node": ">=18.0.0"
+  },
   "scripts": {
-    "dev": "concurrently --names \"app,codegen\" \"npm run gen-type\" \"npm run dev:app\"",
+    "dev": "npm run gen-type && npm run dev:app",
     "dev:app": "vite",
-    "build": "tsc -b && vite build",
-    "lint": "eslint .",
+    "build": "npm run gen-type && tsc -b && vite build",
     "preview": "vite preview",
-    "build:analyze": "vite build --mode analyze",
-    "gen-type": "dotenv -- supabase gen types typescript --project-id=\"$SUPABASE_PROJECT_ID\" > src/types/supabase.ts"
+    "lint": "eslint .",
+    "gen-type": "dotenv -- supabase gen types typescript --project-id=\"$SUPABASE_PROJECT_ID\" > src/shared/types/supabase.ts",
+    "analyze": "vite build --mode analyze"
   },
   "dependencies": {
     "@ant-design/icons": "^6.0.0",
@@ -33,9 +37,9 @@
     "globals": "^16.0.0",
     "rollup-plugin-visualizer": "^5.12.0",
     "supabase": "^2.20.12",
-    "typescript": "~5.8.2",
+    "typescript": "^5.8.2",
     "typescript-eslint": "^8.28.0",
-    "vite": "6.2.6",
-    "concurrently": "^9.1.0"
-  }
+    "vite": "^6.2.6"
+  },
+  "license": "MIT"
 }
