+++ src/app/index/App.css
@@ -80,4 +80,4 @@ h1 {
 .ant-typography {
     text-align: center;
     margin-bottom: 24px !important;
-}
+}
