+++ src/app/index/entry.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { App } from '@/app'
-import './index.css'
+import { App } from './App'
+import '../../index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
         <App/>
     </React.StrictMode>,
-)
+)
