+++ src/app/index/index.ts
@@ -1,0 +1,1 @@
+export { App } from './App';
