+++ src/entities/todo/ui/TodoItem.tsx
@@ -1,7 +1,7 @@
 import {useTodoStore} from '../model/store';
 import {Button, Checkbox, List} from 'antd';
 import {DeleteOutlined} from '@ant-design/icons';
-import {Todo} from "src/entities/todo";
+import {Todo} from "@/entities/todo";
 
 interface TodoItemProps {
     todo: Todo;
