+++ src/shared/config.ts
@@ -1,0 +1,6 @@
+const supabaseUrl: string = import.meta.env.VITE_SUPABASE_URL;
+const supabaseAnonKey: string = import.meta.env.VITE_SUPABASE_ANON_KEY;
+
+export const config = {
+    supabaseUrl,supabaseAnonKey
+}
