--- src/shared/types/generated.ts
@@ -1,184 +1,0 @@
-export type Json =
-    | string
-    | number
-    | boolean
-    | null
-    | { [key: string]: Json | undefined }
-    | Json[]
-
-export type Database = {
-    graphql_public: {
-        Tables: {
-            [_ in never]: never
-        }
-        Views: {
-            [_ in never]: never
-        }
-        Functions: {
-            graphql: {
-                Args: {
-                    operationName?: string
-                    query?: string
-                    variables?: Json
-                    extensions?: Json
-                }
-                Returns: Json
-            }
-        }
-        Enums: {
-            [_ in never]: never
-        }
-        CompositeTypes: {
-            [_ in never]: never
-        }
-    }
-    public: {
-        Tables: {
-            items: {
-                Row: {
-                    completed: boolean
-                    created_at: string
-                    id: number
-                    text: string
-                }
-                Insert: {
-                    completed?: boolean
-                    created_at?: string
-                    id?: number
-                    text: string
-                }
-                Update: {
-                    completed?: boolean
-                    created_at?: string
-                    id?: number
-                    text?: string
-                }
-                Relationships: []
-            }
-            users: {
-                Row: {
-                    created_at: string
-                    id: number
-                }
-                Insert: {
-                    created_at?: string
-                    id?: number
-                }
-                Update: {
-                    created_at?: string
-                    id?: number
-                }
-                Relationships: []
-            }
-        }
-        Views: {
-            [_ in never]: never
-        }
-        Functions: {
-            [_ in never]: never
-        }
-        Enums: {
-            [_ in never]: never
-        }
-        CompositeTypes: {
-            [_ in never]: never
-        }
-    }
-}
-
-type PublicSchema = Database[Extract<keyof Database, "public">]
-
-export type Tables<
-    PublicTableNameOrOptions extends
-            | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
-        | { schema: keyof Database },
-    TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
-        ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
-            Database[PublicTableNameOrOptions["schema"]]["Views"])
-        : never = never,
-> = PublicTableNameOrOptions extends { schema: keyof Database }
-    ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
-        Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
-            Row: infer R
-        }
-        ? R
-        : never
-    : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
-            PublicSchema["Views"])
-        ? (PublicSchema["Tables"] &
-            PublicSchema["Views"])[PublicTableNameOrOptions] extends {
-                Row: infer R
-            }
-            ? R
-            : never
-        : never
-
-export type TablesInsert<
-    PublicTableNameOrOptions extends
-            | keyof PublicSchema["Tables"]
-        | { schema: keyof Database },
-    TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
-        ? keyof Database[PublicTableNameOrOptions["schema"]]["Tables"]
-        : never = never,
-> = PublicTableNameOrOptions extends { schema: keyof Database }
-    ? Database[PublicTableNameOrOptions["schema"]]["Tables"][TableName] extends {
-            Insert: infer I
-        }
-        ? I
-        : never
-    : PublicTableNameOrOptions extends keyof PublicSchema["Tables"]
-        ? PublicSchema["Tables"][PublicTableNameOrOptions] extends {
-                Insert: infer I
-            }
-            ? I
-            : never
-        : never
-
-export type TablesUpdate<
-    PublicTableNameOrOptions extends
-            | keyof PublicSchema["Tables"]
-        | { schema: keyof Database },
-    TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
-        ? keyof Database[PublicTableNameOrOptions["schema"]]["Tables"]
-        : never = never,
-> = PublicTableNameOrOptions extends { schema: keyof Database }
-    ? Database[PublicTableNameOrOptions["schema"]]["Tables"][TableName] extends {
-            Update: infer U
-        }
-        ? U
-        : never
-    : PublicTableNameOrOptions extends keyof PublicSchema["Tables"]
-        ? PublicSchema["Tables"][PublicTableNameOrOptions] extends {
-                Update: infer U
-            }
-            ? U
-            : never
-        : never
-
-export type Enums<
-    PublicEnumNameOrOptions extends
-            | keyof PublicSchema["Enums"]
-        | { schema: keyof Database },
-    EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
-        ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
-        : never = never,
-> = PublicEnumNameOrOptions extends { schema: keyof Database }
-    ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
-    : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
-        ? PublicSchema["Enums"][PublicEnumNameOrOptions]
-        : never
-
-export type CompositeTypes<
-    PublicCompositeTypeNameOrOptions extends
-            | keyof PublicSchema["CompositeTypes"]
-        | { schema: keyof Database },
-    CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
-            schema: keyof Database
-        }
-        ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
-        : never = never,
-> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
-    ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
-    : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
-        ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
-        : never
