+++ tsconfig.json
@@ -21,7 +21,7 @@
     "noFallthroughCasesInSwitch": true,
     "baseUrl": ".",
     "paths": {
-      "src/*": [
+      "@/*": [
         "src/*"
       ]
     }
