+++ vite.config.ts
@@ -7,10 +7,13 @@ export default defineConfig({
     plugins: [react()],
     resolve: {
         alias: {
-            'src': path.resolve(__dirname, './src')
+            '@': path.resolve(__dirname, './src')
         }
     },
     build: {
         sourcemap: true
+    },
+    server: {
+        port: 3000
     }
 })
